Remove fallback cache key
We think we've seen cases where we've seen GHA restore a cache which contains a package installed which has subsequently had an invalid entry added to the poetry lock file. This allows CI to pass with an invalid poetry lock file. If we only restore the cache when the poetry.lock file matches this should avoid this.